Add files via upload
Using the Superset Data, I created a pie chart in Tableau, indicating the total profit for Supersales by region. I first dragged the Region field to the Color area to split the pie chart by regions. After that, I dragged the Profit in the Angle and Label section which shows the per region profit values appropriately. This aids in a visual comparison of what profit in each region contributed. I also added SUM(Sales) to the Marks card, but it’s not used in the final viz – it was part of my earlier attempts. Its a pie chart showing 4 regions profit values on each of its sections for better understandability.